+++ SQLiteStudio3/guiSQLiteStudio/dbtree/dbtree.cpp
@@ -122,6 +122,7 @@ void DbTree::init()
     connect(ui->treeView, SIGNAL(expanded(QModelIndex)), this, SIGNAL(sessionValueChanged()));
     connect(ui->treeView, SIGNAL(collapsed(QModelIndex)), this, SIGNAL(sessionValueChanged()));
 
+    updateIconSize();
     updateActionsForCurrent();
 }
 
@@ -1710,12 +1711,18 @@ void DbTree::changeFontSize(int factor)
     f.setPointSize(f.pointSize() + factor);
     CFG_UI.Fonts.DbTree.set(f);
 
-    QFontMetrics fm(f);
-    ui->treeView->setIconSize(QSize(fm.height(), fm.height()));
-
     f = CFG_UI.Fonts.DbTreeLabel.get();
     f.setPointSize(f.pointSize() + factor);
     CFG_UI.Fonts.DbTreeLabel.set(f);
+
+    updateIconSize();
+}
+
+void DbTree::updateIconSize()
+{
+    auto f = CFG_UI.Fonts.DbTree.get();
+    QFontMetrics fm(f);
+    ui->treeView->setIconSize(QSize(fm.height(), fm.height()));
 }
 
 void DbTree::deleteItems(const QList<DbTreeItem*>& itemsToDelete)