+++ SQLiteStudio3/guiSQLiteStudio/dialogs/configdialog.h
@@ -70,6 +70,7 @@ class GUI_API_EXPORT ConfigDialog : public QDialog
         void initLangs();
         void initTooltips();
         void initColors();
+        void initLookAndFeelPage();
         void applyStyle(QWidget* widget, QStyle* style);
         QTreeWidgetItem* getPluginsCategoryItem() const;
         QTreeWidgetItem* getPluginsCategoryItem(PluginType* type) const;