+++ SQLiteStudio3/guiSQLiteStudio/uiconfig.h
@@ -156,6 +156,7 @@ CFG_UI_CATEGORIES(Ui,
         CFG_ENTRY(bool,                    UseSciFormatForDoubles,      false)
         CFG_ENTRY(bool,                    UseLfForMultilineEditors,    false)
         CFG_ENTRY(Cfg::ColorPickerConfig,  ColorPickerConfig,           Cfg::ColorPickerConfig())
+        CFG_ENTRY(int,                     ToolBarIconSize,             0)
     )
 )